refactor: remove dynamic notch feedback
