+++ Diduny/App/AppDelegate+MeetingRecording.swift
@@ -416,7 +416,7 @@ extension AppDelegate {
         let chapterNumber = appState.meetingChapters.count + 1
         let chapter = MeetingChapter(timestampSeconds: elapsed, label: "Chapter \(chapterNumber)")
         appState.meetingChapters.append(chapter)
-        NotchManager.shared.showInfo(message: "Chapter \(chapterNumber) added", duration: 1.0)
+        DictationOverlayController.shared.showInfo(message: "Chapter \(chapterNumber) added", duration: 1.0)
         Log.app.info("Meeting chapter \(chapterNumber) added at \(elapsed)s")
     }
 
@@ -544,15 +544,15 @@ extension AppDelegate {
                     Task { @MainActor in
                         switch status {
                         case .queued:
-                            NotchManager.shared.showInfo(message: "Queued...", duration: 30)
+                            DictationOverlayController.shared.showInfo(message: "Queued...", duration: 30)
                         case .uploading:
-                            NotchManager.shared.showInfo(message: "Uploading...", duration: 30)
+                            DictationOverlayController.shared.showInfo(message: "Uploading...", duration: 30)
                         case .processing:
                             // Processing can take tens of minutes for large files —
                             // use persistent processing state instead of auto-dismissing info
-                            NotchManager.shared.startProcessing(mode: .meeting)
+                            DictationOverlayController.shared.startProcessing(mode: .meeting)
                         case .finalizing:
-                            NotchManager.shared.showInfo(message: "Finishing up...", duration: 30)
+                            DictationOverlayController.shared.showInfo(message: "Finishing up...", duration: 30)
                         default:
                             break
                         }
@@ -617,7 +617,7 @@ extension AppDelegate {
                 }
 
                 if !cloudModeEnabled {
-                    NotchManager.shared.showInfo(
+                    DictationOverlayController.shared.showInfo(
                         message: "Recording saved. Open Recordings and choose a local model to transcribe.",
                         duration: 3.0
                     )
@@ -727,7 +727,7 @@ extension AppDelegate {
         }
 
         escapeService.onProgressEscape = { pressCount, _ in
-            NotchManager.shared.showInfoDuringRecording(
+            DictationOverlayController.shared.showInfoDuringRecording(
                 message: SettingsStorage.shared.escapeCancelRepeatHint(afterPressCount: pressCount),
                 mode: .meeting,
                 duration: 1.5
@@ -740,7 +740,7 @@ extension AppDelegate {
                 let shouldSaveAudio = SettingsStorage.shared.escapeCancelSaveAudio
                 await self?.cancelMeetingRecording()
                 let message = shouldSaveAudio ? "Recording cancelled and saved" : "Recording cancelled"
-                NotchManager.shared.showInfo(message: message)
+                DictationOverlayController.shared.showInfo(message: message)
             }
         }
