+++ Diduny/App/AppDelegate.swift
@@ -115,7 +115,7 @@ final class AppDelegate: NSObject, NSApplicationDelegate {
 
     let appState = AppState()
 
-    /// Audio level piping to notch
+    /// Audio level piping to the recording feedback panel
     var audioLevelCancellable: AnyCancellable?
 
     // App Nap prevention tokens
@@ -197,7 +197,7 @@ final class AppDelegate: NSObject, NSApplicationDelegate {
         // Permission gate uses AuthService.hasStoredSession (cheap UserDefaults
         // flag) which does not trigger the keychain read.
 
-        setupNotchStopHandler()
+        setupRecordingFeedbackStopHandler()
 
         // Meeting starts fetch SCShareableContent (0.5-2s from the window
         // server); warm the cache now so the first start hits it. No-op until
@@ -315,7 +315,7 @@ final class AppDelegate: NSObject, NSApplicationDelegate {
         Task {
             await RemoteConfigService.shared.fetchIfNeeded()
             if let msg = RemoteConfigService.shared.maintenanceMessage {
-                NotchManager.shared.showInfo(message: msg, duration: 5.0)
+                DictationOverlayController.shared.showInfo(message: msg, duration: 5.0)
             }
         }
 
@@ -390,9 +390,9 @@ final class AppDelegate: NSObject, NSApplicationDelegate {
     }
 
     private func showWakeAfterRecordingInterrupt() {
-        Log.recording.info("[Sleep] wake after recording interrupt — surfacing notch message")
+        Log.recording.info("[Sleep] wake after recording interrupt — surfacing feedback message")
 
-        NotchManager.shared.showInfo(
+        DictationOverlayController.shared.showInfo(
             message: "Recording stopped. Open Recordings to recover audio.",
             duration: 5.0
         )
@@ -507,12 +507,12 @@ final class AppDelegate: NSObject, NSApplicationDelegate {
                 Log.app.error("Recovery transcription failed: \(error.localizedDescription)")
                 if let recoveredRecordingID {
                     MainWindowController.shared.showRecording(id: recoveredRecordingID)
-                    NotchManager.shared.showInfo(
+                    DictationOverlayController.shared.showInfo(
                         message: "Recording recovered. Transcription can be retried from Recordings.",
                         duration: 5
                     )
                 } else {
-                    NotchManager.shared.showInfo(
+                    DictationOverlayController.shared.showInfo(
                         message: "Recovery failed. The original audio was kept for another attempt.",
                         duration: 5
                     )
@@ -577,7 +577,7 @@ final class AppDelegate: NSObject, NSApplicationDelegate {
                         duration: 2.0
                     )
                 } else {
-                    NotchManager.shared.showInfo(message: "Microphone disconnected", duration: 2.0)
+                    DictationOverlayController.shared.showInfo(message: "Microphone disconnected", duration: 2.0)
                 }
             }
         }
@@ -585,16 +585,13 @@ final class AppDelegate: NSObject, NSApplicationDelegate {
 
     // MARK: - Cross-Mode Recording Guard
 
-    private func setupNotchStopHandler() {
-        NotchManager.shared.setStopHandler { [weak self] in
-            await self?.stopActiveRecordingFromNotch()
-        }
+    private func setupRecordingFeedbackStopHandler() {
         DictationOverlayController.shared.setStopHandler { [weak self] in
-            await self?.stopActiveRecordingFromNotch()
+            await self?.stopActiveRecordingFromFeedback()
         }
     }
 
-    func stopActiveRecordingFromNotch() async {
+    func stopActiveRecordingFromFeedback() async {
         if appState.meetingTranslationRecordingState == .recording {
             await stopMeetingTranslationRecording()
             return
@@ -635,14 +632,14 @@ final class AppDelegate: NSObject, NSApplicationDelegate {
             return
         }
 
-        Log.app.info("stopActiveRecordingFromNotch: no active recording state")
+        Log.app.info("stopActiveRecordingFromFeedback: no active recording state")
     }
 
     private func isStateInProgress(_ state: RecordingState) -> Bool {
         state == .recording || state == .processing
     }
 
-    private func restoreNotchForActiveRecordingAfterInfo(delay: TimeInterval = 1.6) {
+    private func restoreRecordingFeedbackAfterInfo(delay: TimeInterval = 1.6) {
         Task { @MainActor [weak self] in
             try? await Task.sleep(for: .seconds(delay))
             guard let self else { return }
@@ -688,8 +685,8 @@ final class AppDelegate: NSObject, NSApplicationDelegate {
     func wireMeetingRecorderStatusMessages() {
         meetingRecorderService.onStatusMessage = { [weak self] message in
             Task { @MainActor in
-                NotchManager.shared.showInfo(message: message, duration: 2.0)
-                self?.restoreNotchForActiveRecordingAfterInfo(delay: 2.1)
+                DictationOverlayController.shared.showInfo(message: message, duration: 2.0)
+                self?.restoreRecordingFeedbackAfterInfo(delay: 2.1)
             }
         }
 
@@ -710,7 +707,7 @@ final class AppDelegate: NSObject, NSApplicationDelegate {
                     self.handleMeetingStateChange(.error)
                 }
 
-                NotchManager.shared.showError(message: error.localizedDescription)
+                DictationOverlayController.shared.showError(message: error.localizedDescription)
             }
         }
     }
@@ -755,9 +752,9 @@ final class AppDelegate: NSObject, NSApplicationDelegate {
                 duration: 1.5
             )
         } else {
-            NotchManager.shared.showInfo(message: "Stop current recording first", duration: 1.5)
+            DictationOverlayController.shared.showInfo(message: "Stop current recording first", duration: 1.5)
         }
-        restoreNotchForActiveRecordingAfterInfo()
+        restoreRecordingFeedbackAfterInfo()
 
         return false
     }
